Making linear solvers independent libraries #40
