+++ Project.toml
@@ -5,7 +5,6 @@ repo = "https://github.com/sshin23/MadNLP.jl.git"
 version = "0.2.0"
 
 [deps]
-IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
@@ -18,19 +17,31 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SuiteSparse = "4607b0f0-06f3-5cda-b6b1-a6196a1729e9"
 
 [compat]
-IterativeSolvers = "~0.9.1"
 MathOptInterface = "~0.9"
 NLPModels = "~0.14, ~0.15"
 SolverCore = "~0.1"
 julia = "~1"
-
-# AmplNLReader = "~0.9"
-# JuMP = "~0.21"
-# Plasmo = "~0.3, ~0.4"
+JuMP = "~0.21"
+Plasmo = "~0.3, ~0.4"
+AmplNLReader = "~0.9"
+MadNLPGPU = "~0.1" 
+MadNLPHSL = "~0.1" 
+MadNLPIterative = "~0.1" 
+MadNLPMumps = "~0.1" 
+MadNLPPardiso = "~0.1" 
+MadNLPGraphs = "~0.1"
 
 [extras]
-AmplNLReader = "77dd3d4c-cb1d-5e09-9340-85030ff7ba66"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+Plasmo = "d3f7391f-f14a-50cc-bbe4-76a32d1bad3c"
+AmplNLReader = "77dd3d4c-cb1d-5e09-9340-85030ff7ba66"
+MadNLPGPU = "d72a61cc-809d-412f-99be-fd81f4b8a598"
+MadNLPHSL = "7fb6135f-58fe-4112-84ca-653cf5be0c77"
+MadNLPIterative = "1888cb03-ce40-4a36-8d45-ae8231a0e17c"
+MadNLPMumps = "3b83494e-c0a4-4895-918b-9157a7a085a1"
+MadNLPPardiso = "312ee924-cb12-49df-b284-7304c3902fc0"
+MadNLPGraphs = "2f590f30-ea46-438a-a78d-e5bfe3109704"
 
 [targets]
-test = ["Test", "AmplNLReader"]
+test = ["Test", "JuMP", "Plasmo", "AmplNLReader", "MadNLPGPU", "MadNLPHSL", "MadNLPIterative", "MadNLPMumps", "MadNLPPardiso", "MadNLPGraphs"]
