+++ lib/MadNLPHSL/Project.toml
@@ -8,11 +8,9 @@ MadNLP = "2621e9c9-9eb4-46b1-8089-e8c72242dfb6"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 BinaryProvider = "b99e7846-7c00-51b0-8f62-c81ae34c0232"
 OpenBLAS32_jll = "656ef2d0-ae68-5445-9ca0-591084a874a2"
-MKL_jll = "856f044c-d86e-5d09-b602-aeab76dc8ba7"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
-julia = "~1"
 
 [compat]
 OpenBLAS32_jll = "~0.3"
-MKL_jll = "~2020"
 MadNLP = "~0.2"
+julia = "~1"
