+++ lib/MadNLPHSL/deps/build.jl
@@ -1,15 +1,7 @@
 # MadNLPHSL.jl
 # Created by Sungho Shin (sungho.shin@wisc.edu)
 
-using Pkg.Artifacts, BinaryProvider, OpenBLAS32_jll, MKL_jll
-
-# if haskey(ENV,"MADNLP_BLAS") # TODO: MKL support 
-#     blasvendor = ENV["MADNLP_BLAS"]=="openblas" ? :openblas : :mkl
-# else
-#     blasvendor = MKL_jll.best_platform == nothing ? :openblas : :mkl
-# end
-
-blasvendor = :openblas
+using Pkg.Artifacts, BinaryProvider, OpenBLAS32_jll
 
 const verbose = "--verbose" in ARGS
 const prefix = Prefix(@__DIR__)
@@ -29,8 +21,6 @@ const optimization_flag = haskey(ENV,"MADNLP_OPTIMIZATION_FLAG") ? ENV["MADNLP_O
 const installer = Sys.isapple() ? "brew install" : Sys.iswindows() ? "pacman -S" : "sudo apt install"
 const libopenblas_dir = splitdir(OpenBLAS32_jll.libopenblas_path)[1]
 const with_openblas = `-L$libopenblas_dir $rpath$libopenblas_dir -lopenblas`
-const libmkl_dir = joinpath(MKL_jll.artifact_dir,MKL_jll.libmkl_rt_splitpath[1:end-1]...)
-const with_mkl = `-L$libmkl_dir $rpath$libmkl_dir -lmkl_rt`
 
 rm.(filter(endswith(".so"), readdir(libdir,join=true)))
 products   = Product[]
@@ -63,7 +53,7 @@ if hsl_shared_library == ""
             push!(OC,OutputCollector(`$FC $openmp_flag -fPIC -c $optimization_flag -o hsl_mc68/C/hsl_mc68i_ciface.o hsl_mc68/C/hsl_mc68i_ciface.f90`,verbose=verbose))
             push!(OC,OutputCollector(`$FC $openmp_flag -fPIC -c $optimization_flag -o hsl_ma97/C/hsl_ma97d_ciface.o hsl_ma97/C/hsl_ma97d_ciface.f90`,verbose=verbose))
             wait.(OC); empty!(OC)
-            wait(OutputCollector(`$FC -o$(libdir)/libhsl.$so -shared -fPIC $optimization_flag common/deps90.o common/deps.o mc19/mc19d.o ma27/ma27d.o ma57/ma57d.o hsl_ma77/hsl_ma77d.o hsl_ma77/C/hsl_ma77d_ciface.o hsl_ma86/hsl_ma86d.o hsl_ma86/C/hsl_ma86d_ciface.o hsl_mc68/C/hsl_mc68i_ciface.o hsl_ma97/hsl_ma97d.o hsl_ma97/C/hsl_ma97d_ciface.o $openmp_flag $with_metis $(blasvendor == :mkl ? with_mkl : with_openblas)`,verbose=verbose))
+            wait(OutputCollector(`$FC -o$(libdir)/libhsl.$so -shared -fPIC $optimization_flag common/deps90.o common/deps.o mc19/mc19d.o ma27/ma27d.o ma57/ma57d.o hsl_ma77/hsl_ma77d.o hsl_ma77/C/hsl_ma77d_ciface.o hsl_ma86/hsl_ma86d.o hsl_ma86/C/hsl_ma86d_ciface.o hsl_mc68/C/hsl_mc68i_ciface.o hsl_ma97/hsl_ma97d.o hsl_ma97/C/hsl_ma97d_ciface.o $openmp_flag $with_metis $with_openblas`,verbose=verbose))
             cd("$(@__DIR__)")
         end
         @info "Building HSL (Ma27, Ma57, Ma77, Ma86, Ma97) $(build_succeded(products[end]))."
