+++ lib/MadNLPIterative/Project.toml
@@ -1,0 +1,13 @@
+name = "MadNLPIterative"
+uuid = "1888cb03-ce40-4a36-8d45-ae8231a0e17c"
+authors = ["Sungho Shin <sungho.shin.ss@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
+MadNLP = "2621e9c9-9eb4-46b1-8089-e8c72242dfb6"
+
+[compat]
+IterativeSolvers = "~0.9"
+MadNLP = "~0.2"
+julia = "~1"
