+++ lib/MadNLPPardiso/Project.toml
@@ -13,3 +13,4 @@ Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 [compat]
 OpenBLAS32_jll = "~0.3"
 MadNLP = "~0.2"
+julia = "~1"
