+++ src/utils.jl
@@ -2,8 +2,9 @@
 # Created by Sungho Shin (sungho.shin@wisc.edu)
 
 # Build info
-default_linear_solver() = Umfpack
-default_dense_solver() = LapackCPU
+default_linear_solver() = MadNLPUmfpack
+default_dense_solver() = MadNLPLapackCPU
+default_iterator() = MadNLPRichardson
 
 # Options
 abstract type AbstractOptions end
